requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
amqp==5.2.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.35.52
botocore==1.35.52
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-cleanup==9.0.0
django-filter==24.3
django-jalali-date==1.1.3
django-storages==1.14.4
django-widget-tweaks==1.5.0
dnspython==2.7.0
env==0.1.0
eventlet==0.37.0
exceptiongroup==1.2.2
ffmpeg-python==0.2.0
fontawesome-free==5.15.4
fonttools==4.55.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jalali_core==1.0.0
jdatetime==5.0.0
jmespath==1.0.1
kombu==5.4.2
multidict==6.1.0
phonenumbers==8.13.52
pillow==10.4.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.1.0
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.22
pydyf==0.11.0
pyOpenSSL==24.2.1
PyPDF2==3.0.1
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.7
redis==5.2.0
requests==2.32.3
retrying==1.3.4
rsa==4.9
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
Telethon==1.38.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
yarl==1.17.1
zopfli==0.2.3.post1
